// src/app/interfaces/skill/skill.enums.ts
// Enums de skills y categorías (const enum)
export const enum SkillDescription {
  ANGULAR = "Angular",
  BASH = "Bash",
  BOOTSTRAP = "Bootstrap",
  BULMA = "Bulma",
  CSS3 = "CSS3",
  DATABASES = "SQL Server",
  DOCKER = "Docker",
  ECLIPSE = "Eclipse",
  GIT = "Git",
  GITHUB = "GitHub",
  HTML5 = "HTML5",
  JAVA = "Java",
  JAVASCRIPT = "JavaScript",
  LINUX = "Linux",
  NGINX = "Nginx",
  NODE = "Node.js",
  POWERSHELL = "PowerShell",
  PYTHON = "Python",
  TAILWIND = "Tailwind",
  TYPESCRIPT = "TypeScript",
  VISUAL = "Visual Basic (net)"
}
export const enum SkillIcon {
  ANGULAR = "angular.svg",
  BASH = "bash.svg",
  BOOTSTRAP = "bootstrap.svg",
  BULMA = "bulma.svg",
  CSS3 = "css3.svg",
  DATABASES = "database.svg",
  DOCKER = "docker.svg",
  ECLIPSE = "eclipse.svg",
  GIT = "git.svg",
  GITHUB = "github.svg",
  HTML5 = "html5.svg",
  JAVA = "java.svg",
  JAVASCRIPT = "javascript.svg",
  LINUX = "linux.svg",
  NGINX = "nginx.svg",
  NODE = "node.svg",
  POWERSHELL = "powershell.svg",
  PYTHON = "python.svg",
  TAILWIND = "tailwind.svg",
  TYPESCRIPT = "typescript.svg",
  VISUAL = "visual.svg"
}
export const enum SkillLink {
  ANGULAR = "https://angular.io/",
  BASH = "https://www.gnu.org/software/bash/",
  BOOTSTRAP = "https://getbootstrap.com/",
  BULMA = "https://bulma.io/",
  CSS3 = "https://www.w3.org/Style/CSS/Overview.en.html",
  DATABASES = "https://www.microsoft.com/en-us/sql-server/sql-server-downloads",
  DOCKER = "https://www.docker.com/",
  ECLIPSE = "https://www.eclipse.org/",
  GIT = "https://git-scm.com/",
  GITHUB = "https://github.com/JarolRojas",
  HTML5 = "https://developer.mozilla.org/en-US/docs/Web/Guide/HTML/HTML5",
  JAVA = "https://www.java.com/",
  JAVASCRIPT = "https://developer.mozilla.org/en-US/docs/Web/JavaScript",
  LINUX = "https://www.linux.org/",
  NGINX = "https://nginx.org/",
  NODE = "https://nodejs.org/",
  POWERSHELL = "https://docs.microsoft.com/en-us/powershell/",
  PYTHON = "https://www.python.org/",
  TAILWIND = "https://tailwindcss.com/",
  TYPESCRIPT = "https://www.typescriptlang.org/",
  VISUAL = "https://learn.microsoft.com/en-us/dotnet/visual-basic/"
}
export const enum SkillColor {
  ANGULAR = "#dd0031",
  BASH = "#4eaa25",
  BOOTSTRAP = "#7952b3",
  BULMA = "#00d1b2",
  CSS3 = "#264de4",
  DATABASES = "#00758f",
  DOCKER = "#2496ed",
  ECLIPSE = "#2c2255",
  GIT = "#f34f29",
  GITHUB = "#181717",
  HTML5 = "#e34c26",
  JAVA = "#007396",
  JAVASCRIPT = "#f7df1e",
  LINUX = "#fcc624",
  NGINX = "#009639",
  NODE = "#339933",
  POWERSHELL = "#012456",
  PYTHON = "#3776ab",
  TAILWIND = "#38bdf8",
  TYPESCRIPT = "#3178c6",
  VISUAL = "#68217a"
}



export const enum CategoryType {
  FRONTEND = "FRONTEND",
  BACKEND = "BACKEND",
  DEVOPS = "DEVOPS",
  DATABASE = "DATABASE",
  OTHER = "OTHER"
}
export const enum CategoryIcon {
  FRONTEND = "fa-solid fa-code",
  BACKEND = "fa-solid fa-server",
  DEVOPS = "fa-solid fa-cloud",
  DATABASE = "fa-solid fa-database",
  OTHER = "fa-solid fa-ellipsis-h"
}